pyproject.toml
[project]
name = "trp1-week3-document-intelligence-refinery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.2",
    "docling>=2.76.0",
    "fitz>=0.0.1.dev2",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langdetect>=1.0.9",
    "langgraph>=1.0.10",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "tiktoken>=0.12.0",
    "typer>=0.21.2",
]
